Moved callbacks to file; Added callback to task_hpo and removed manual saving;
